Install shellcheck only if needed

+++ bootstrap-ubuntu.sh
@@ -126,14 +126,18 @@ sudo chown "$USER": /usr/local/evm
 curl -fsSkL https://raw.github.com/rejeep/evm/master/go | bash
 
 echo "Install ShellCheck..."
-mkdir -p ~/tmp && cd ~/tmp || exit 1
-wget --quiet "https://storage.googleapis.com/shellcheck/shellcheck-stable.linux.x86_64.tar.xz"
-tar --xz -xvf shellcheck-stable.linux.x86_64.tar.xz
-cp shellcheck-stable/shellcheck ~/bin/shellcheck
-rm -rf shellcheck-stable
-rm -f shellcheck-stable.linux.x86_64.tar.xz
+mkdir -p ~/bin
+if [ ! -f ~/bin/pup ]; then
+  mkdir -p ~/tmp && cd ~/tmp || exit 1
+  wget --quiet "https://storage.googleapis.com/shellcheck/shellcheck-stable.linux.x86_64.tar.xz"
+  tar --xz -xvf shellcheck-stable.linux.x86_64.tar.xz
+  cp shellcheck-stable/shellcheck ~/bin/shellcheck
+  rm -rf shellcheck-stable
+  rm -f shellcheck-stable.linux.x86_64.tar.xz
+fi
 
 echo "Install pup..."
+mkdir -p ~/bin
 if [ ! -f ~/bin/pup ]; then
   cd ~/tmp || exit 1
   PUP_VERSION=0.4.0
